+++ src/@locale/demo/scheduler/index.js
@@ -1,12 +1,14 @@
 import scheduleModel   from './scheduleModel' ;
 import schedulerAgenda from './schedulerAgenda' ;
 import schedulerMonth  from './schedulerMonth' ;
+import schedulerPalette from './schedulerPalette' ;
 
 const scheduler =
 {
     scheduleModel ,
     schedulerAgenda ,
     schedulerMonth ,
+    schedulerPalette ,
 } ;
 
 export default scheduler ;