+++ src/components/charts/MarimekkoChart.jsx
@@ -15,7 +15,7 @@ import { useMedia } from 'react-use' ;
 import isMarimekkoDataValid from '../../helpers/charts/isMarimekkoDataValid' ;
 
 import useChartLayout  from '../../hooks/useChartLayout' ;
-import useChartPalette from '../../hooks/useChartPalette' ;
+import usePalette from '../../hooks/usePalette' ;
 import useChartTheme   from '../../hooks/useChartTheme' ;
 
 import { CARTESIAN } from '../../themes/charts/layout' ;
@@ -117,7 +117,7 @@ const MarimekkoChart =
 {
     const theme = useChartTheme( { overrides : themeOverrides } ) ;
 
-    const colors = useChartPalette( { palette , count : dimensions?.length ?? 0 } ) ;
+    const colors = usePalette( { palette , count : dimensions?.length ?? 0 } ) ;
 
     const reduceMotion = useMedia( '(prefers-reduced-motion: reduce)' , false ) ;